requirements.txt
flask==2.3.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flask_sqlalchemy==3.1.1
flask_mail==0.9.1
gunicorn==21.2.0